rearranged the code and the files
